ran black on all files

+++ foolbox/tests/batch_attacks/test_batch_attacks_bapp.py
@@ -6,8 +6,7 @@ from foolbox.distances import Linf
 
 def test_attack(bn_model, bn_criterion, bn_images, bn_labels):
     attack = BoundaryAttackPlusPlus(bn_model, bn_criterion)
-    advs = attack(bn_images, bn_labels, unpack=False,
-                  iterations=20, verbose=True)
+    advs = attack(bn_images, bn_labels, unpack=False, iterations=20, verbose=True)
     for adv in advs:
         assert adv.perturbed is not None
         assert adv.distance.value < np.inf
@@ -15,8 +14,7 @@ def test_attack(bn_model, bn_criterion, bn_images, bn_labels):
 
 def test_attack_linf(bn_model, bn_criterion, bn_images, bn_labels):
     attack = BoundaryAttackPlusPlus(bn_model, bn_criterion, distance=Linf)
-    advs = attack(bn_images, bn_labels, unpack=False,
-                  iterations=20, verbose=True)
+    advs = attack(bn_images, bn_labels, unpack=False, iterations=20, verbose=True)
     for adv in advs:
         assert adv.perturbed is not None
         assert adv.distance.value < np.inf
@@ -24,8 +22,7 @@ def test_attack_linf(bn_model, bn_criterion, bn_images, bn_labels):
 
 def test_attack_non_verbose(bn_model, bn_criterion, bn_images, bn_labels):
     attack = BoundaryAttackPlusPlus(bn_model, bn_criterion)
-    advs = attack(bn_images, bn_labels, unpack=False,
-                  iterations=20, verbose=False)
+    advs = attack(bn_images, bn_labels, unpack=False, iterations=20, verbose=False)
     for adv in advs:
         assert adv.perturbed is not None
         assert adv.distance.value < np.inf
@@ -48,20 +45,25 @@ def test_attack_continue(bn_adversarial):
 
 def test_attack_targeted(bn_model, bn_criterion, bn_images, bn_labels):
     np.random.seed(2)
-    starting_point = np.random.uniform(
-        0, 1, size=bn_images[0].shape).astype(bn_images.dtype)
+    starting_point = np.random.uniform(0, 1, size=bn_images[0].shape).astype(
+        bn_images.dtype
+    )
 
     attack = BoundaryAttackPlusPlus(bn_model, bn_criterion)
-    advs = attack(bn_images, bn_labels, unpack=False,
-                  iterations=21,
-                  starting_point=starting_point,
-                  log_every_n_steps=2,
-                  gamma=0.01,
-                  stepsize_search='geometric_progression',
-                  batch_size=128,
-                  initial_num_evals=200,
-                  max_num_evals=20000,
-                  verbose=True)
+    advs = attack(
+        bn_images,
+        bn_labels,
+        unpack=False,
+        iterations=21,
+        starting_point=starting_point,
+        log_every_n_steps=2,
+        gamma=0.01,
+        stepsize_search="geometric_progression",
+        batch_size=128,
+        initial_num_evals=200,
+        max_num_evals=20000,
+        verbose=True,
+    )
     for adv in advs:
         assert adv.perturbed is not None
         assert adv.distance.value < np.inf
@@ -69,20 +71,25 @@ def test_attack_targeted(bn_model, bn_criterion, bn_images, bn_labels):
 
 def test_attack_linf_targeted(bn_model, bn_criterion, bn_images, bn_labels):
     np.random.seed(2)
-    starting_point = np.random.uniform(
-        0, 1, size=bn_images[0].shape).astype(bn_images.dtype)
+    starting_point = np.random.uniform(0, 1, size=bn_images[0].shape).astype(
+        bn_images.dtype
+    )
 
     attack = BoundaryAttackPlusPlus(bn_model, bn_criterion, distance=Linf)
-    advs = attack(bn_images, bn_labels, unpack=False,
-                  iterations=21,
-                  starting_point=starting_point,
-                  log_every_n_steps=2,
-                  gamma=0.01,
-                  stepsize_search='grid_search',
-                  batch_size=128,
-                  initial_num_evals=200,
-                  max_num_evals=20000,
-                  verbose=True)
+    advs = attack(
+        bn_images,
+        bn_labels,
+        unpack=False,
+        iterations=21,
+        starting_point=starting_point,
+        log_every_n_steps=2,
+        gamma=0.01,
+        stepsize_search="grid_search",
+        batch_size=128,
+        initial_num_evals=200,
+        max_num_evals=20000,
+        verbose=True,
+    )
     for adv in advs:
         assert adv.perturbed is not None
         assert adv.distance.value < np.inf
@@ -90,18 +97,15 @@ def test_attack_linf_targeted(bn_model, bn_criterion, bn_images, bn_labels):
 
 def test_attack_gl(gl_bn_model, bn_criterion, bn_images, bn_labels):
     attack = BoundaryAttackPlusPlus(gl_bn_model, bn_criterion)
-    advs = attack(bn_images, bn_labels, unpack=False,
-                  iterations=20, verbose=True)
+    advs = attack(bn_images, bn_labels, unpack=False, iterations=20, verbose=True)
     for adv in advs:
         assert adv.perturbed is not None
         assert adv.distance.value < np.inf
 
 
-def test_attack_impossible(bn_model, bn_impossible_criterion, bn_images,
-                           bn_labels):
+def test_attack_impossible(bn_model, bn_impossible_criterion, bn_images, bn_labels):
     attack = BoundaryAttackPlusPlus(bn_model, bn_impossible_criterion)
-    advs = attack(bn_images, bn_labels, unpack=False,
-                  iterations=20, verbose=True)
+    advs = attack(bn_images, bn_labels, unpack=False, iterations=20, verbose=True)
     for adv in advs:
         assert adv.perturbed is None
         assert adv.distance.value == np.inf
